+++ firmware/Components/RTD/Inc/rtd.h
@@ -26,8 +26,8 @@ extern volatile uint16_t RTD_ADC_REGISTER[RTD_SAMPLES];
 /* Public function prototypes ------------------------------------------------*/
 /**
  * @brief Read the temperature reading from the RTD sensor
- * @retval The temperature reading in degrees Celsius
+ * @retval The temperature reading in degrees Celsius * 100
  */
-float RTD_GetTemperature(void);
+int16_t RTD_GetTemperature(void);
 
 #endif /* __RTD_H__ */
